Controle de segurança para deletar itens de venda.

+++ src/main/java/br/com/caracore/pdv/service/ItemVendaService.java
@@ -25,7 +25,9 @@ public class ItemVendaService {
 	}
 	
 	public void excluir(Long codigo) {
-		itemVendaRepository.delete(codigo);
+		if (Util.validar(codigo)) {
+			itemVendaRepository.delete(codigo);
+		}
 	}
 
 	List<ItemVenda> buscarItens(Venda venda) {
